Reorganized 3d model folders, changed antenna pegs to square.
Reorganized 3D models into one folder, with added documentation, to support a mockup assembly.

Replaced previous three round pegs with a single square peg with an alignment peg attached to one side. Should allow for a single top cap that can be attached in four orientations for the 4 antenna array.